Separate VertexState into separate compilation unit
