+++ CMakeLists.txt
@@ -877,6 +877,7 @@ else()
     src/glview/Renderer.cc
     src/glview/system-gl.cc
     src/glview/VertexArray.cc
+    src/glview/VertexState.cc
     src/glview/VBORenderer.cc
     src/glview/GLView.cc
     src/glview/VertexStateManager.cc
